Core/Fog/Filter/requirements.txt
paho-mqtt==1.3.1
influxdb==5.0.0